requirements.txt
haystack-ai
python-dotenv
pandas
SQLAlchemy
pymysql

torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
numpy>=1.24.0